requirements.txt
requests~=2.24.0
Pillow~=7.2.0
ImageHash~=4.1.0